+++ src/engine/audio_music.cpp
@@ -56,7 +56,7 @@ void Music::Play( const std::vector<u8> & v, bool loop )
 {
     if ( Mixer::isValid() && v.size() ) {
         // u32 id = CheckSum( v );
-        SDL_RWops * rwops = SDL_RWFromConstMem( &v[0], v.size() );
+        SDL_RWops * rwops = SDL_RWFromConstMem( &v[0], static_cast<int>(v.size()) );
 #if SDL_VERSION_ATLEAST( 2, 0, 0 )
         Mix_Music * mix = Mix_LoadMUS_RW( rwops, 0 );
 #else
